+++ lib/src/inspector.dart
@@ -1,10 +1,12 @@
 import 'dart:ui' as ui;
 
 import 'package:flutter/foundation.dart';
+import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
 import 'package:inspector/src/keyboard_handler.dart';
+import 'package:inspector/src/widgets/zoom/zoom_overlay.dart';
 
 import './widgets/panel/inspector_panel.dart';
 import 'utils.dart';
@@ -46,6 +48,7 @@ class Inspector extends StatefulWidget {
     this.isPanelVisible = true,
     this.isWidgetInspectorEnabled = true,
     this.isColorPickerEnabled = true,
+    this.isZoomEnabled = true,
     this.widgetInspectorShortcuts = const [
       LogicalKeyboardKey.alt,
       LogicalKeyboardKey.altLeft,
@@ -59,6 +62,9 @@ class Inspector extends StatefulWidget {
       LogicalKeyboardKey.shiftLeft,
       LogicalKeyboardKey.shiftRight,
     ],
+    this.zoomShortcuts = const [
+      LogicalKeyboardKey.keyZ,
+    ],
     this.isEnabled,
   }) : super(key: key);
 
@@ -67,9 +73,11 @@ class Inspector extends StatefulWidget {
   final bool isPanelVisible;
   final bool isWidgetInspectorEnabled;
   final bool isColorPickerEnabled;
+  final bool isZoomEnabled;
   final Alignment alignment;
   final List<LogicalKeyboardKey> widgetInspectorShortcuts;
   final List<LogicalKeyboardKey> colorPickerShortcuts;
+  final List<LogicalKeyboardKey> zoomShortcuts;
   final bool? isEnabled;
 
   @override
@@ -88,11 +96,18 @@ class _InspectorState extends State<Inspector> {
 
   final _inspectorStateNotifier = ValueNotifier<bool>(false);
   final _colorPickerStateNotifier = ValueNotifier<bool>(false);
+  final _zoomStateNotifier = ValueNotifier<bool>(false);
 
   final _selectedColorOffsetNotifier = ValueNotifier<Offset?>(null);
   final _selectedColorStateNotifier = ValueNotifier<Color?>(null);
 
+  final _zoomImageOffsetNotifier = ValueNotifier<Offset?>(null);
+  final _zoomScaleNotifier = ValueNotifier<double>(2.0);
+  final _zoomOverlayOffsetNotifier = ValueNotifier<Offset?>(null);
+
   late final KeyboardHandler _keyboardHandler;
+
+  Offset? _pointerHoverPosition;
 
   @override
   void initState() {
@@ -101,8 +116,10 @@ class _InspectorState extends State<Inspector> {
     _keyboardHandler = KeyboardHandler(
       onInspectorStateChanged: _onInspectorStateChanged,
       onColorPickerStateChanged: _onColorPickerStateChanged,
+      onZoomStateChanged: _onZoomStateChanged,
       colorPickerStateKeys: widget.colorPickerShortcuts,
       inspectorStateKeys: widget.widgetInspectorShortcuts,
+      zoomStateKeys: widget.zoomShortcuts,
     );
 
     if (_isEnabled && widget.areKeyboardShortcutsEnabled) {
@@ -115,10 +132,11 @@ class _InspectorState extends State<Inspector> {
   void _onTap(Offset? pointerOffset) {
     if (_colorPickerStateNotifier.value) {
       if (pointerOffset != null) {
-        _onHover(pointerOffset);
+        _onColorPickerHover(pointerOffset);
       }
 
       _onColorPickerStateChanged(false);
+      return;
     }
 
     if (!_inspectorStateNotifier.value) {
@@ -146,7 +164,14 @@ class _InspectorState extends State<Inspector> {
 
   void _onPointerMove(Offset pointerOffset) {
     if (_colorPickerStateNotifier.value) {
-      _onHover(pointerOffset);
+      _onColorPickerHover(pointerOffset);
+    }
+  }
+
+  void _onPointerHover(Offset pointerOffset) {
+    _pointerHoverPosition = pointerOffset;
+    if (_zoomStateNotifier.value) {
+      _onZoomHover(pointerOffset);
     }
   }
 
@@ -162,6 +187,7 @@ class _InspectorState extends State<Inspector> {
 
     if (isEnabled) {
       _onColorPickerStateChanged(false);
+      _onZoomStateChanged(false);
     } else {
       _currentRenderBoxNotifier.value = null;
     }
@@ -179,6 +205,8 @@ class _InspectorState extends State<Inspector> {
 
     if (isEnabled) {
       _onInspectorStateChanged(false);
+      _onZoomStateChanged(false);
+
       WidgetsBinding.instance.addPostFrameCallback((_) {
         _extractByteData();
       });
@@ -199,21 +227,66 @@ class _InspectorState extends State<Inspector> {
     }
   }
 
+  // Zoom
+
+  void _onZoomStateChanged(bool isEnabled) {
+    if (!widget.isZoomEnabled) {
+      _zoomStateNotifier.value = false;
+      return;
+    }
+
+    _zoomStateNotifier.value = isEnabled;
+
+    if (isEnabled) {
+      _onInspectorStateChanged(false);
+      _onColorPickerStateChanged(false);
+      _zoomScaleNotifier.value = 2.0;
+
+      WidgetsBinding.instance.addPostFrameCallback((_) async {
+        await _extractByteData();
+
+        if (_pointerHoverPosition != null) {
+          _onZoomHover(_pointerHoverPosition!);
+        }
+      });
+    } else {
+      _image?.dispose();
+      _image = null;
+      _byteDataStateNotifier.value = null;
+
+      _zoomImageOffsetNotifier.value = null;
+      _zoomOverlayOffsetNotifier.value = null;
+      _zoomScaleNotifier.value = 2.0;
+    }
+  }
+
   Future<void> _extractByteData() async {
+    if (_image != null) return;
     final boundary = _repaintBoundaryKey.currentContext!.findRenderObject()!
         as RenderRepaintBoundary;
 
-    _image = await boundary.toImage();
+    final pixelRatio = MediaQuery.of(context).devicePixelRatio;
+
+    _image = await boundary.toImage(pixelRatio: pixelRatio);
     _byteDataStateNotifier.value = await _image!.toByteData();
   }
 
-  void _onHover(Offset offset) {
-    if (_image == null || _byteDataStateNotifier.value == null) return;
+  Offset _extractShiftedOffset(Offset offset) {
+    final pixelRatio = MediaQuery.of(context).devicePixelRatio;
 
-    final shiftedOffset = (_repaintBoundaryKey.currentContext!
-            .findRenderObject()! as RenderRepaintBoundary)
+    var _offset = (_repaintBoundaryKey.currentContext!.findRenderObject()!
+            as RenderRepaintBoundary)
         .globalToLocal(offset);
 
+    _offset *= pixelRatio;
+
+    return _offset;
+  }
+
+  void _onColorPickerHover(Offset offset) {
+    if (_image == null || _byteDataStateNotifier.value == null) return;
+
+    final shiftedOffset = _extractShiftedOffset(offset);
     final _x = shiftedOffset.dx.round();
     final _y = shiftedOffset.dy.round();
 
@@ -229,6 +302,32 @@ class _InspectorState extends State<Inspector> {
             .localToGlobal(Offset.zero);
 
     _selectedColorOffsetNotifier.value = offset - overlayOffset;
+  }
+
+  void _onZoomHover(Offset offset) {
+    if (_image == null || _byteDataStateNotifier.value == null) return;
+
+    final shiftedOffset = _extractShiftedOffset(offset);
+
+    final overlayOffset =
+        (_stackKey.currentContext!.findRenderObject() as RenderStack)
+            .localToGlobal(Offset.zero);
+
+    _zoomImageOffsetNotifier.value = shiftedOffset;
+    _zoomOverlayOffsetNotifier.value = offset - overlayOffset;
+  }
+
+  void _onPointerScroll(PointerScrollEvent scrollEvent) {
+    if (_zoomStateNotifier.value) {
+      final newValue =
+          _zoomScaleNotifier.value + 1.0 * -scrollEvent.scrollDelta.dy.sign;
+
+      if (newValue < 1.0) {
+        return;
+      }
+
+      _zoomScaleNotifier.value = newValue;
+    }
   }
 
   @override
@@ -274,31 +373,34 @@ class _InspectorState extends State<Inspector> {
             valueListenables: [
               _colorPickerStateNotifier,
               _inspectorStateNotifier,
+              _zoomStateNotifier,
             ],
             builder: (context) {
               Widget _child = widget.child;
 
-              if (_colorPickerStateNotifier.value ||
-                  _inspectorStateNotifier.value) {
-                _child = AbsorbPointer(
-                  key: _absorbPointerKey,
-                  child: _child,
-                );
-              }
-
-              if (_colorPickerStateNotifier.value) {
-                _child = RepaintBoundary(
-                  key: _repaintBoundaryKey,
-                  child: _child,
-                );
-              }
+              final isAbsorbingPointer = _colorPickerStateNotifier.value ||
+                  _inspectorStateNotifier.value ||
+                  _zoomStateNotifier.value;
 
               return Listener(
                 behavior: HitTestBehavior.translucent,
                 onPointerUp: (e) => _onTap(e.position),
                 onPointerMove: (e) => _onPointerMove(e.position),
                 onPointerDown: (e) => _onPointerMove(e.position),
-                child: _child,
+                onPointerHover: (e) => _onPointerHover(e.position),
+                onPointerSignal: (event) {
+                  if (event is PointerScrollEvent) {
+                    _onPointerScroll(event);
+                  }
+                },
+                child: RepaintBoundary(
+                  key: _repaintBoundaryKey,
+                  child: AbsorbPointer(
+                    key: _absorbPointerKey,
+                    absorbing: isAbsorbingPointer,
+                    child: _child,
+                  ),
+                ),
               );
             },
           ),
@@ -331,6 +433,7 @@ class _InspectorState extends State<Inspector> {
             valueListenables: [
               _currentRenderBoxNotifier,
               _inspectorStateNotifier,
+              _zoomStateNotifier,
             ],
             builder: (context) => LayoutBuilder(
               builder: (context, constraints) => _inspectorStateNotifier.value
@@ -341,6 +444,48 @@ class _InspectorState extends State<Inspector> {
                   : const SizedBox.shrink(),
             ),
           ),
+        if (widget.isZoomEnabled)
+          MultiValueListenableBuilder(
+            valueListenables: [
+              _zoomImageOffsetNotifier,
+              _zoomOverlayOffsetNotifier,
+              _byteDataStateNotifier,
+              _zoomScaleNotifier,
+            ],
+            builder: (context) {
+              final offset = _zoomOverlayOffsetNotifier.value;
+              final imageOffset = _zoomImageOffsetNotifier.value;
+              final byteData = _byteDataStateNotifier.value;
+              final zoomScale = _zoomScaleNotifier.value;
+
+              if (offset == null || byteData == null || imageOffset == null) {
+                return const SizedBox.shrink();
+              }
+
+              final overlaySize = ui
+                  .lerpDouble(
+                    128.0,
+                    256.0,
+                    ((zoomScale - 2.0) / 10.0).clamp(0, 1),
+                  )!
+                  .toDouble();
+
+              return Positioned(
+                left: offset.dx - overlaySize / 2,
+                top: offset.dy - overlaySize / 2,
+                child: IgnorePointer(
+                  child: ZoomOverlayWidget(
+                    image: _image!,
+                    overlayOffset: offset,
+                    imageOffset: imageOffset,
+                    overlaySize: overlaySize,
+                    zoomScale: zoomScale,
+                    pixelRatio: MediaQuery.of(context).devicePixelRatio,
+                  ),
+                ),
+              );
+            },
+          ),
         if (widget.isPanelVisible)
           Align(
             alignment: Alignment.centerRight,
@@ -348,15 +493,20 @@ class _InspectorState extends State<Inspector> {
               valueListenables: [
                 _inspectorStateNotifier,
                 _colorPickerStateNotifier,
+                _zoomStateNotifier,
                 _byteDataStateNotifier,
               ],
               builder: (context) => InspectorPanel(
                 isInspectorEnabled: _inspectorStateNotifier.value,
                 isColorPickerEnabled: _colorPickerStateNotifier.value,
+                isZoomEnabled: _zoomStateNotifier.value,
                 onInspectorStateChanged: _onInspectorStateChanged,
                 onColorPickerStateChanged: _onColorPickerStateChanged,
+                onZoomStateChanged: _onZoomStateChanged,
                 isColorPickerLoading: _byteDataStateNotifier.value == null &&
                     _colorPickerStateNotifier.value,
+                isZoomLoading: _byteDataStateNotifier.value == null &&
+                    _zoomStateNotifier.value,
               ),
             ),
           ),